+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * Rename `button.new` from "+ New" to "New". Refs STCOR-1093.
 * Supply `rtrIgnore` on forgot-password and forgot-username API requests. Refs STCOR-1099.
 * Codesmell cleanup in ForgotUserName, ForgotPassword. Refs STCOR-993.
+* Given an invalid locale, default to `en-US` and recover. Refs STCOR-1102.
 * Support notifications plugin in the main navigation. Refs STCOR-1094.
 
 ## [11.1.1](https://github.com/folio-org/stripes-core/tree/v11.1.1) (2026-04-20)